Multiple changes to DogtailTestCase
- remove anaconda logs before starting a new test
- update the environment if the test provides one
- kill anaconda after test execution and remove anaconda pid
- don't run if no tests defined
- rename __init__.py to base.py so that nose doesn't run tests
  from the base class and update all the imports
- save individual test results into separate directories
