+++ tests/gui/keyboard.py
@@ -15,7 +15,7 @@
 #
 # Author: Chris Lumens <clumens@redhat.com>
 
-from . import UITestCase
+from .base import UITestCase
 
 class BasicKeyboardTestCase(UITestCase):
     def check_options_dialog(self, spoke):
