+++ tests/gui/storage.py
@@ -18,7 +18,7 @@
 from dogtail.predicate import GenericPredicate
 from dogtail.utils import doDelay
 
-from . import UITestCase
+from .base import UITestCase
 
 class BasicStorageTestCase(UITestCase):
     def check_select_disks(self, spoke):
